# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mermaid Validator API",
    "version": "1.4.0",
    "description": "Validates Mermaid diagrams (~32 types) using vendored jison + langium grammars.\n\nLive list of supported types: `GET /api/v1/capabilities`\n\n### Result is tri-state\n\n| `valid` | `status` | meaning |\n|---|---|---|\n| `true` | `validated` | parsed, content OK |\n| `false` | `invalid` | parsed, content has errors |\n| `null` | `unsupported` | no parser — we could not check |\n\nGate on `valid !== true` to keep \"do not ship\" behavior for both `false` and `null`.\n\n### Endpoint stability\n\n- Validators (`/validate`, `/upload/file`, `/markdown/validate`, `/health/*`, `/stats`, `/capabilities`): stable.\n- Auto-fixers (`/markdown/fix`, `/upload/fix`): **BETA** — heuristic rewriter, review the diff before accepting.\n\n### Notes\n\n- `%%` line-comments are stripped before parsing.\n- `%%{init: ...}%%` theme directives are preserved.\n- Grammars are vendored snapshots, not a live wrapper around upstream `mermaid`.",
    "contact": {
      "name": "Gregorio Elias Roecker Momm",
      "email": "gregoriomomm@gmail.com"
    },
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error"
        ]
      },
      "DiagramInput": {
        "type": "object",
        "description": "A single diagram to validate",
        "properties": {
          "id": {
            "type": "string",
            "example": "diagram_1"
          },
          "content": {
            "type": "string",
            "example": "flowchart TD\n  A-->B"
          },
          "type": {
            "type": "string",
            "example": "flowchart",
            "nullable": true
          }
        },
        "required": [
          "content"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "syntax_error"
          },
          "message": {
            "type": "string"
          },
          "line": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "DiagramResult": {
        "type": "object",
        "description": "Per-diagram validation result. The `valid` field is tri-state: true = parsed cleanly; false = parsed with errors; null = the validator could not run (unknown type or no parser). The `status` field gives the same information in string form.",
        "properties": {
          "id": {
            "type": "string"
          },
          "valid": {
            "type": "boolean",
            "nullable": true,
            "description": "true=ok, false=invalid, null=could not validate"
          },
          "status": {
            "type": "string",
            "enum": [
              "validated",
              "invalid",
              "unsupported"
            ],
            "description": "Matches the tri-state valid field"
          },
          "diagramType": {
            "type": "string",
            "nullable": true,
            "description": "Detected diagram type (null if undetectable)"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "MarkdownValidateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "totalDiagrams": {
            "type": "integer"
          },
          "validDiagrams": {
            "type": "integer"
          },
          "invalidDiagrams": {
            "type": "integer"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagramResult"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/validate": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Validate one or more Mermaid diagrams (JSON body)",
        "description": "Direct validation against the custom Jison/Langium grammar parser. No file upload required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "diagrams"
                ],
                "properties": {
                  "diagrams": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DiagramInput"
                    }
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "timeout": {
                        "type": "integer",
                        "description": "Per-diagram timeout in ms"
                      }
                    }
                  }
                }
              },
              "examples": {
                "flowchart": {
                  "summary": "One flowchart",
                  "value": {
                    "diagrams": [
                      {
                        "id": "diagram_1",
                        "content": "flowchart TD\n  A-->B"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "processingTime": {
                      "type": "integer"
                    },
                    "validator": {
                      "type": "string",
                      "example": "custom_grammar_parser"
                    },
                    "totalDiagrams": {
                      "type": "integer"
                    },
                    "validDiagrams": {
                      "type": "integer"
                    },
                    "invalidDiagrams": {
                      "type": "integer"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DiagramResult"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad input (e.g.",
            "too many diagrams)": null,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/markdown/fix": {
      "post": {
        "tags": [
          "markdown"
        ],
        "summary": "[BETA] Auto-fix all Mermaid diagrams inside a markdown document",
        "description": "**BETA — heuristic auto-fixer.** Extracts every fenced ```mermaid block, attempts iterative auto-fix (up to 5 passes per diagram), and returns the rewritten markdown plus statistics.\n\nThe fixer applies pattern-based corrections (common arrow typos, missing keywords, malformed brackets). It is **not** a semantic rewriter — it can change diagrams in subtle ways and does NOT guarantee the fixed output matches your intent. Treat as a best-effort transform and review the diff before accepting.\n\nStability: experimental. The fix patterns and response shape may change between minor versions.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Full markdown body"
                  },
                  "options": {
                    "type": "object",
                    "description": "Optional fixer overrides",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Markdown processed (success may be false if some diagrams could not be fixed)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "fixedContent": {
                      "type": "string"
                    },
                    "statistics": {
                      "type": "object",
                      "properties": {
                        "totalDiagrams": {
                          "type": "integer"
                        },
                        "fixedDiagrams": {
                          "type": "integer"
                        },
                        "failedDiagrams": {
                          "type": "integer"
                        },
                        "totalIterations": {
                          "type": "integer"
                        },
                        "processingTime": {
                          "type": "integer"
                        }
                      }
                    },
                    "diagrams": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "success": {
                            "type": "boolean"
                          },
                          "wasFixed": {
                            "type": "boolean"
                          },
                          "iterations": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "report": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/markdown/validate": {
      "post": {
        "tags": [
          "markdown"
        ],
        "summary": "Validate (no fixing) all Mermaid diagrams in a markdown document",
        "description": "Extracts every fenced ```mermaid block and validates each. Does not modify content.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Full markdown body"
                  }
                }
              },
              "examples": {
                "tiny": {
                  "summary": "One inline flowchart",
                  "value": {
                    "content": "```mermaid\nflowchart TD\n  A-->B\n```"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation completed (success=true iff invalidDiagrams===0)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkdownValidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}